create a fucking window
Honestly it's the first time that I have trouble to create not even a
"hello world" like program. The tutorial was meant for webgl and wasm so
I wasted almost an hour to figure out that this won't work IDK why.

So finally I found a challenge... This is going to be a funny learning
time.